+++ configs/pixart/inference/1x512x512.py
@@ -24,7 +24,7 @@ scheduler = dict(
     num_sampling_steps=20,
     cfg_scale=7.0,
 )
-dtype = "fp16"
+dtype = "bf16"
 
 # prompt_path = "./assets/texts/t2i_samples.txt"
 prompt = [
